Basic code to call local function from lib
This isn't a hook per se, but it uses the same basic mechanism. I've
reversed the very simple libtarget.so binary (just using objdump) and
located the local function `say_something`.

By calculating its offset from the start of the `maybe_say` function,
which is a global symbol whose address can be determined at runtime
using `dlsym`, we can create a function pointer to it.