Basic "hook" for calling private member function
As with the C example, this is not truly a hook, but it does call the
private member function without needing to modify the library's code by
changing the scope for that function.

This turns out to be slightly easier than the C example, because member
functions in (non-local) C++ classes always have external linkage, so
there's no chance of a missing symbol, and you can use `dlsym` with the
mangled function name to get the address directly.