Complex lib target for private member fn hooking
This is slightly more complex than the basic target, because the private
member function to be hooked / called makes use of `this`.

That means an actual instance of the class must be constructed and setup
according to the calling convention of the private function before the
call can be made.