Example of "hooking" the complex target lib
This example is slightly more complex than the basic example because we
need to construct an instance of the Target class first. A pointer to
the instance must also be placed on the stack (or in correct register)
according to the calling convention of the private member function.

After that setup has been completed, the call is actually the same as
the basic and C examples, just calling the function pointer.