Call private member function w/o dlsym()
This isn't really anything close to hooking -- the symbol->address
resolution is being done by the dynamic linker entirely -- but it's
certainly the simplest option.

If you can modify your code's version of the class declaration so that
the private member function you wish to call is actually public, then
the linker will make it work using "normal" C++ code.